syncnet_python/requirements.txt
anyio==3.7.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cycler==0.11.0
decorator==4.4.2
exceptiongroup==1.2.2
fastapi==0.95.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.10
imageio==2.31.2
imageio-ffmpeg==0.5.1
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.3.4
numpy==1.18.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
Pillow==9.5.0
platformdirs==4.0.0
proglog==0.1.10
pydantic==1.10.19
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
python-multipart==0.0.8
python-speech-features==0.6
PyYAML==6.0.1
requests==2.31.0
scenedetect==0.6.4
scipy==1.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
zipp==3.15.0
